Damage modifier is a float now I guess
- Luke

+++ DrBounce-UnityProject/Assets/Scripts/Gun.cs
@@ -6,7 +6,7 @@ using UnityEngine;
 public class Gun : ScriptableObject
 {
     public int baseDamage;    //default damage of the gun 
-    public int damageModifier;    //multiplied with the amount of bounces 
+    public float damageModifier;    //multiplied with the amount of bounces 
     public int amountOfChargesGiven;    //amount of charges given when you rebound the gun
     public float fireRate;    //time between shots
 }
